Added selectable damageType to Towers
Added a dropdown menu to select damagetype on tower prefab.. get passed to the bullet. and minion know what it get hits by (shows in debug at the moment)

+++ Tower Defense Game/Assets/Assets Mads/MinionSubunits/Intel.cs
@@ -6,7 +6,7 @@ public class Intel : MonoBehaviour
 {
     public bool reportIntel = false;
     public string intel;
-    private int health;
+    private float health;
     
     void Update()
     {
